pyproject.toml
[project]
name = "charts-downloader"
version = "0.1.0"
description = "Stock charts downloader and viewer"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask",
    "playwright",
    "python-dotenv",
    "Pillow",
]

[tool.uv]
dev-dependencies = []
